Move custom editor code to extra folder
