+++ src/meson.build
@@ -6,10 +6,11 @@ configure_file(input : 'config.h.in',
                output : 'config.h',
                configuration : conf_data)
 
-razergenie_sources = ['main.cpp', 'razergenie.cpp', 'razerimagedownloader.cpp', 'razerdevicewidget.cpp', 'devicelistwidget.cpp', 'customeditor.cpp', 'matrixpushbutton.cpp', 'util.cpp']
+razergenie_sources = ['main.cpp', 'razergenie.cpp', 'razerimagedownloader.cpp', 'razerdevicewidget.cpp', 'devicelistwidget.cpp', 'util.cpp',
+                      'customeditor/customeditor.cpp', 'customeditor/matrixpushbutton.cpp']
 
 processed = qt5.preprocess(
-  moc_headers : ['razergenie.h', 'razerimagedownloader.h', 'customeditor.h', 'devicelistwidget.h'],
+  moc_headers : ['razergenie.h', 'razerimagedownloader.h', 'customeditor/customeditor.h', 'devicelistwidget.h'],
   ui_files : '../ui/razergenie.ui'
 )
